restructuring package-server-script module
