+++ test/unit/cli-domain-package-server-script.js
@@ -19,7 +19,7 @@ var initialise = function(fs){
     writeFile: sinon.stub().yields(null, 'ok')
   }, fs || {});
 
-  packageServerScript = injectr('../../src/cli/domain/package-server-script.js', {
+  packageServerScript = injectr('../../src/cli/domain/package-server-script/index.js', {
     'fs-extra': fsMock,
     path: {
       extname: path.extname,
